refactor: move cache_langs to utils
this is the right way to ensure  can always run
