+++ g2p/mappings/langs/utils.py
@@ -4,6 +4,15 @@ Utilities used by other classes
 
 """
 
+from pathlib import Path
+
+import yaml
+import pickle
+from networkx import DiGraph, write_gpickle
+
+from g2p.mappings.langs import LANGS_DIR, LANGS_PKL, LANGS_NWORK_PATH
+from g2p.mappings.utils import load_mapping_from_path
+
 from g2p.log import LOGGER
 from g2p.mappings import Mapping
 from g2p.mappings.langs import MAPPINGS_AVAILABLE
@@ -117,3 +126,56 @@ def is_arpabet(string):
         if sound not in _ARPABET_SET:
             return False
     return True
+
+
+def cache_langs(dir_path: str = LANGS_DIR,
+                langs_path: str = LANGS_PKL,
+                network_path: str = LANGS_NWORK_PATH):
+    """Read in all files and save as pickle.
+
+    Args:
+       dir_path: Path to scan for config.yaml files.  Default is the
+                 installed g2p/mappings/langs directory.
+       langs_path: Path to output langs.pkl pickle file.  Default is
+                   the installed g2p/mappings/langs/langs.pkl
+       network_path: Path to output pickle file.  Default is the
+                     installed g2p/mappings/langs/network.pkl.
+    """
+    langs = {}
+    dir_path = Path(dir_path)
+
+    # Sort by language code
+    paths = sorted(dir_path.glob("./*/config.y*ml"),
+                   key=lambda x: x.parent.stem)
+    mappings_legal_pairs = []
+    for path in paths:
+        code = path.parent.stem
+        with open(path, encoding="utf8") as f:
+            data = yaml.safe_load(f)
+        # If there is a mappings key, there is more than one mapping
+        # TODO: should put in some measure to prioritize non-generated
+        # mappings and warn when they override
+        if "mappings" in data:
+            for index, mapping in enumerate(data["mappings"]):
+                mappings_legal_pairs.append(
+                    (
+                        data["mappings"][index]["in_lang"],
+                        data["mappings"][index]["out_lang"],
+                    )
+                )
+                data["mappings"][index] = load_mapping_from_path(path, index)
+        else:
+            data = load_mapping_from_path(path)
+        langs[code] = data
+
+    # Save as a Directional Graph
+    lang_network = DiGraph()
+    lang_network.add_edges_from(mappings_legal_pairs)
+
+    with open(network_path, "wb") as f:
+        write_gpickle(lang_network, f, protocol=4)
+
+    with open(langs_path, "wb") as f:
+        pickle.dump(langs, f, protocol=4)
+
+    return langs
